LayoutTests/js/symbol-abstract-relational-comparison-expected.txt

This tests Abstract Relatioal Comparison results with Symbols.

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".


PASS 42 < symbol threw exception TypeError: Type error.
PASS symbol < 42 threw exception TypeError: Type error.
PASS NaN < symbol threw exception TypeError: Type error.
PASS symbol < NaN threw exception TypeError: Type error.
PASS Infinity < symbol threw exception TypeError: Type error.
PASS symbol < Infinity threw exception TypeError: Type error.
PASS true < symbol threw exception TypeError: Type error.
PASS symbol < true threw exception TypeError: Type error.
PASS false < symbol threw exception TypeError: Type error.
PASS symbol < false threw exception TypeError: Type error.
PASS null < symbol threw exception TypeError: Type error.
PASS symbol < null threw exception TypeError: Type error.
PASS undefined < symbol threw exception TypeError: Type error.
PASS symbol < undefined threw exception TypeError: Type error.
PASS 'Cappuccino' < symbol threw exception TypeError: Type error.
PASS symbol < 'Cappuccino' threw exception TypeError: Type error.
PASS symbol < symbol threw exception TypeError: Type error.
PASS symbol < symbol threw exception TypeError: Type error.
PASS Symbol.iterator < symbol threw exception TypeError: Type error.
PASS symbol < Symbol.iterator threw exception TypeError: Type error.
PASS object < symbol threw exception TypeError: Type error.
PASS symbol < object threw exception TypeError: Type error.
PASS array < symbol threw exception TypeError: Type error.
PASS symbol < array threw exception TypeError: Type error.
PASS date < symbol threw exception TypeError: Type error.
PASS symbol < date threw exception TypeError: Type error.
PASS 42 <= symbol threw exception TypeError: Type error.
PASS symbol <= 42 threw exception TypeError: Type error.
PASS NaN <= symbol threw exception TypeError: Type error.
PASS symbol <= NaN threw exception TypeError: Type error.
PASS Infinity <= symbol threw exception TypeError: Type error.
PASS symbol <= Infinity threw exception TypeError: Type error.
PASS true <= symbol threw exception TypeError: Type error.
PASS symbol <= true threw exception TypeError: Type error.
PASS false <= symbol threw exception TypeError: Type error.
PASS symbol <= false threw exception TypeError: Type error.
PASS null <= symbol threw exception TypeError: Type error.
PASS symbol <= null threw exception TypeError: Type error.
PASS undefined <= symbol threw exception TypeError: Type error.
PASS symbol <= undefined threw exception TypeError: Type error.
PASS 'Cappuccino' <= symbol threw exception TypeError: Type error.
PASS symbol <= 'Cappuccino' threw exception TypeError: Type error.
PASS symbol <= symbol threw exception TypeError: Type error.
PASS symbol <= symbol threw exception TypeError: Type error.
PASS Symbol.iterator <= symbol threw exception TypeError: Type error.
PASS symbol <= Symbol.iterator threw exception TypeError: Type error.
PASS object <= symbol threw exception TypeError: Type error.
PASS symbol <= object threw exception TypeError: Type error.
PASS array <= symbol threw exception TypeError: Type error.
PASS symbol <= array threw exception TypeError: Type error.
PASS date <= symbol threw exception TypeError: Type error.
PASS symbol <= date threw exception TypeError: Type error.
PASS 42 > symbol threw exception TypeError: Type error.
PASS symbol > 42 threw exception TypeError: Type error.
PASS NaN > symbol threw exception TypeError: Type error.
PASS symbol > NaN threw exception TypeError: Type error.
PASS Infinity > symbol threw exception TypeError: Type error.
PASS symbol > Infinity threw exception TypeError: Type error.
PASS true > symbol threw exception TypeError: Type error.
PASS symbol > true threw exception TypeError: Type error.
PASS false > symbol threw exception TypeError: Type error.
PASS symbol > false threw exception TypeError: Type error.
PASS null > symbol threw exception TypeError: Type error.
PASS symbol > null threw exception TypeError: Type error.
PASS undefined > symbol threw exception TypeError: Type error.
PASS symbol > undefined threw exception TypeError: Type error.
PASS 'Cappuccino' > symbol threw exception TypeError: Type error.
PASS symbol > 'Cappuccino' threw exception TypeError: Type error.
PASS symbol > symbol threw exception TypeError: Type error.
PASS symbol > symbol threw exception TypeError: Type error.
PASS Symbol.iterator > symbol threw exception TypeError: Type error.
PASS symbol > Symbol.iterator threw exception TypeError: Type error.
PASS object > symbol threw exception TypeError: Type error.
PASS symbol > object threw exception TypeError: Type error.
PASS array > symbol threw exception TypeError: Type error.
PASS symbol > array threw exception TypeError: Type error.
PASS date > symbol threw exception TypeError: Type error.
PASS symbol > date threw exception TypeError: Type error.
PASS 42 >= symbol threw exception TypeError: Type error.
PASS symbol >= 42 threw exception TypeError: Type error.
PASS NaN >= symbol threw exception TypeError: Type error.
PASS symbol >= NaN threw exception TypeError: Type error.
PASS Infinity >= symbol threw exception TypeError: Type error.
PASS symbol >= Infinity threw exception TypeError: Type error.
PASS true >= symbol threw exception TypeError: Type error.
PASS symbol >= true threw exception TypeError: Type error.
PASS false >= symbol threw exception TypeError: Type error.
PASS symbol >= false threw exception TypeError: Type error.
PASS null >= symbol threw exception TypeError: Type error.
PASS symbol >= null threw exception TypeError: Type error.
PASS undefined >= symbol threw exception TypeError: Type error.
PASS symbol >= undefined threw exception TypeError: Type error.
PASS 'Cappuccino' >= symbol threw exception TypeError: Type error.
PASS symbol >= 'Cappuccino' threw exception TypeError: Type error.
PASS symbol >= symbol threw exception TypeError: Type error.
PASS symbol >= symbol threw exception TypeError: Type error.
PASS Symbol.iterator >= symbol threw exception TypeError: Type error.
PASS symbol >= Symbol.iterator threw exception TypeError: Type error.
PASS object >= symbol threw exception TypeError: Type error.
PASS symbol >= object threw exception TypeError: Type error.
PASS array >= symbol threw exception TypeError: Type error.
PASS symbol >= array threw exception TypeError: Type error.
PASS date >= symbol threw exception TypeError: Type error.
PASS symbol >= date threw exception TypeError: Type error.
PASS successfullyParsed is true

TEST COMPLETE
